move `ResultsViewController` helpers to a separate file
